// src/payment/types.ts
export interface CreatePortalParams {
  customerId: string;
  returnUrl?: string;
  locale?: Locale;
}